# source/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/role/add": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/delete": {
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/get-all": {
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Role_Search"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/model.Role_View"
                                    }
                                },
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/get-one": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "$ref": "#/definitions/model.Role_View"
                                },
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/role/update": {
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "role"
                ],
                "parameters": [
                    {
                        "description": "PARAM",
                        "name": "parameter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Role"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "meta_data": {
                                    "$ref": "#/definitions/model.MetadataResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.MetadataResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/model.PaginationResponse"
                },
                "status": {
                    "type": "boolean"
                },
                "timeExecution": {
                    "type": "string"
                }
            }
        },
        "model.PaginationResponse": {
            "type": "object",
            "properties": {
                "size": {
                    "type": "integer"
                },
                "totalElements": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
        "model.Range": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer",
                    "example": 1646792565000
                },
                "field": {
                    "type": "string",
                    "example": "updatedAt"
                },
                "start": {
                    "type": "integer",
                    "example": 1646792565000
                }
            }
        },
        "model.Role": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "privileges": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "updatedAt": {
                    "type": "integer"
                }
            }
        },
        "model.Role_Search": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "order": {
                    "type": "string",
                    "example": "DESC"
                },
                "orderBy": {
                    "type": "string",
                    "example": "createdAt"
                },
                "page": {
                    "type": "integer",
                    "example": 1
                },
                "range": {
                    "$ref": "#/definitions/model.Range"
                },
                "search": {
                    "description": "? Regex",
                    "type": "string"
                },
                "searchBy": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "size": {
                    "type": "integer",
                    "example": 11
                }
            }
        },
        "model.Role_View": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "privileges": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": true
                    }
                },
                "updatedAt": {
                    "type": "integer"
                },
                "userCount": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "JWT": {
            "description": "E.g. Bearer Your.Token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}